requirements.txt
Django==1.11.4
django-cors-headers==2.1.0
djangorestframework==3.6.4
pkg-resources==0.0.0
pytz==2017.2
